+++ semver/comparator_test.ts
@@ -4,8 +4,6 @@ import { assertEquals } from "../assert/mod.ts";
 import { parseRange } from "./parse_range.ts";
 import { parse } from "./parse.ts";
 import { testRange } from "./test_range.ts";
-import { parseComparator } from "./parse_comparator.ts";
-import { comparatorFormat } from "./comparator_format.ts";
 
 Deno.test({
   name: "comparators",
@@ -159,15 +157,4 @@ Deno.test({
       });
     }
   },
-});
-
-Deno.test("tostrings", function () {
-  assertEquals(
-    comparatorFormat(parseComparator(">= v1.2.3")),
-    ">=1.2.3",
-  );
-  assertEquals(
-    comparatorFormat(parseComparator(">= v1.2.3-pre.1+b.2")),
-    ">=1.2.3-pre.1+b.2",
-  );
 });
